requirements.txt
flask==3.1.0
twilio==9.4.1
gspread==6.1.4
google-auth==2.38.0
google-generativeai==0.8.4
gunicorn==23.0.0
